+++ resources/views/test.blade.php
@@ -10,13 +10,3 @@
 
 
 @endsection
-
-
-
-    <div id="one">
-        @{{ shared.user.name }}
-    </div>
-
-    <div id="two">
-        @{{ shared.user.name }}
-    </div>
